re-import the main pkgdist module back into its canonical repo
It doesn't really belong in snakeoil.dist since it has to be bundled
everywhere else at the moment to properly bootstrap pip installs.

If the bootstrapping issue can ever be cleanly solved (i.e. forcing
snakeoil to be installed before setup.py is parsed via pip or whatever)
then it probably would make sense to move it back to snakeoil or
somewhere else.